Specifications

+++ src/main/java/entity/Cone.java
@@ -1,5 +1,5 @@
 package entity;
 
-public record Cone(int id, Point center, double radius, double height) {
+public record Cone(long id, Point center, double radius, double height) {
 
 }
